Parse commands successfully

+++ strutil.h
@@ -1,7 +1,7 @@
 #ifndef ORGACOMPU_TP2_STRUTIL_H
 #define ORGACOMPU_TP2_STRUTIL_H
 
-char** split(const char* str, char sep, size_t* size);
+char** split(const char* str, char sep, int * size);
 void free_split(char **strv);
 
 #endif //ORGACOMPU_TP2_STRUTIL_H
